requirements.txt
requests==2.31.0
python-dotenv==1.0.0
typer==0.9.0
rich==13.7.0
pydantic==2.5.3
